requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.1.2
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.111.1
fastapi-cli==0.0.16
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.19.2
PyMySQL==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.16.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typer==0.20.0
typing_extensions==4.15.0
uvicorn==0.23.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
stripe>=7.0.0
